moved to proptree due to hyphen_undersorce issue with setuptools. Now with working setuptools
