+++ .gitignore
@@ -1,10 +1,9 @@
 __pycache__/
 build/
-
+dist/
+.vscode/
 *.py[cod]
-
 *.so
-
 *.json
 *.xml
 *.ini
