+++ pyproject.toml
@@ -1,0 +1,21 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+[project]
+name = "proptree"
+version = "1.0"
+authors = [
+  {name = "Dan Eicher" },
+]
+description='Python bindings for C++ boost::property_tree'
+readme = "README.md"
+requires-python = ">=3.7"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: Boost Software License 1.0 (BSL-1.0)",    
+    "Operating System :: OS Independent",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/eponymous/python3-property_tree"
+"Bug Tracker" = "https://https://github.com/eponymous/python3-property_tree"
